Extract composables from PostRsListActivity into screens package
Move PostRsListContent and TabContent composables into a new
PostRsListScreen.kt in the screens package, keeping the activity
focused on wiring and navigation.
